+++ React_Master_04_02/src/styles.css
@@ -1,3 +1,73 @@
+/** {*/
+/*    box-sizing: border-box;*/
+/*}*/
+
+/*body {*/
+/*    font-family: sans-serif;*/
+/*    margin: 20px;*/
+/*    padding: 0;*/
+/*}*/
+
+/*h1 {*/
+/*    margin-top: 0;*/
+/*    font-size: 22px;*/
+/*}*/
+
+/*h2 {*/
+/*    margin-top: 0;*/
+/*    font-size: 20px;*/
+/*}*/
+
+/*h3 {*/
+/*    margin-top: 0;*/
+/*    font-size: 18px;*/
+/*}*/
+
+/*h4 {*/
+/*    margin-top: 0;*/
+/*    font-size: 16px;*/
+/*}*/
+
+/*h5 {*/
+/*    margin-top: 0;*/
+/*    font-size: 14px;*/
+/*}*/
+
+/*h6 {*/
+/*    margin-top: 0;*/
+/*    font-size: 12px;*/
+/*}*/
+
+/*code {*/
+/*    font-size: 1.2em;*/
+/*}*/
+
+/*ul {*/
+/*    padding-left: 20px;*/
+/*}*/
+
+/*h2 {*/
+/*    margin-top: 10px;*/
+/*    margin-bottom: 0;*/
+/*}*/
+
+/*h3 {*/
+/*    margin-top: 5px;*/
+/*    font-weight: normal;*/
+/*    font-size: 100%;*/
+/*}*/
+
+/*img {*/
+/*    width: 120px;*/
+/*    height: 120px;*/
+/*}*/
+
+/*button {*/
+/*    display: block;*/
+/*    margin-top: 10px;*/
+/*    margin-bottom: 10px;*/
+/*}*/
+
 * {
     box-sizing: border-box;
 }
@@ -44,6 +114,17 @@ code {
 
 ul {
     padding-left: 20px;
+}
+
+button {
+    display: block;
+    margin-bottom: 10px;
+}
+
+.Page > * {
+    float: left;
+    width: 50%;
+    padding: 10px;
 }
 
 h2 {
